requirement.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
assemblyai==0.29.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
distro==1.9.0
Django==5.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
openai==1.35.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytube==15.0.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
websockets==12.0
Werkzeug==3.0.3
youtube-dl==2021.12.17
